==> 01_Setup_Dependencies.ipynb <==
# Setup All Workshop Dependencies

## _Note:  This Notebook Will Take A Few Minutes To Complete._

## _Please Be Patient._

!python --version

!pip list

# Pip

!pip install --disable-pip-version-check -q pip --upgrade > /dev/null
!pip install --disable-pip-version-check -q wrapt --upgrade > /dev/null

### _Ignore any warning or error message ^^ above ^^.  This is OK!_

# AWS CLI and AWS Python SDK (boto3)

!pip install --disable-pip-version-check -q awscli boto3

### _Ignore any warning or error message ^^ above ^^.  This is OK!_

# SageMaker

!pip install --disable-pip-version-check -q sagemaker
!pip install --disable-pip-version-check -q smdebug
!pip install --disable-pip-version-check -q sagemaker-experiments

### _Ignore any warning or error message ^^ above ^^.  This is OK!_

# PyAthena

!pip install --disable-pip-version-check -q PyAthena

### _Ignore any warning or error message ^^ above ^^.  This is OK!_

# AWS Data Wrangler

!pip install --disable-pip-version-check -q awswrangler

### _Ignore any warning or error message ^^ above ^^.  This is OK!_

# Zip

!conda install -y zip

# Matplotlib

!pip install --disable-pip-version-check -q matplotlib

### _Ignore any warning or error message ^^ above ^^.  This is OK!_

# Seaborn

!pip install --disable-pip-version-check -q seaborn

### _Ignore any warning or error message ^^ above ^^.  This is OK!_

# Summarize

!python --version

!pip list

setup_dependencies_passed = True

%store setup_dependencies_passed

%store

# Release Resources

%%html

<p><b>Shutting down your kernel for this notebook to release resources.</b></p>
<button class="sm-command-button" data-commandlinker-command="kernelmenu:shutdown" style="display:none;">Shutdown Kernel</button>
        
<script>
try {
    els = document.getElementsByClassName("sm-command-button");
    els[0].click();
}
catch(err) {
    // NoOp
}    
</script>

%%javascript

try {
    Jupyter.notebook.save_checkpoint();
    Jupyter.notebook.session.delete();
}
catch(err) {
    // NoOp
}
