==> 03_Copy_data_to_S3.ipynb <==
## Copy data to S3

%store -r setup_dependencies_passed

try:
    setup_dependencies_passed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup Dependencies.")
    print("+++++++++++++++++++++++++++++++")

print(setup_dependencies_passed)

%store -r setup_s3_bucket_passed

try:
    setup_s3_bucket_passed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup S3 Bucket.")
    print("+++++++++++++++++++++++++++++++")

print(setup_s3_bucket_passed)

if not setup_dependencies_passed:
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup Dependencies.")
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
if not setup_s3_bucket_passed:
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup S3 Bucket.")
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")

import boto3
import sagemaker
import pandas as pd

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name
account_id = boto3.client("sts").get_caller_identity().get("Account")

sm = boto3.Session().client(service_name="sagemaker", region_name=region)

# Define the path to your dataset in S3
s3_private_path_csv = "s3://{}/root/AAI-540_Predictive-Maintenance-for-Pharmaceutical-Manufacturing-Equipment/resampled_predictive_maintenance_dataset.csv".format(bucket)

# Print the path to verify
print(s3_private_path_csv)

%store s3_private_path_csv

s3_private_path_csv = "s3://{}/root/AAI-540_Predictive-Maintenance-for-Pharmaceutical-Manufacturing-Equipment/resampled_predictive_maintenance_dataset.csv".format(bucket)


!aws s3 cp "resampled_predictive_maintenance_dataset.csv" $s3_private_path_csv


import boto3
import sagemaker
import pandas as pd
import io

# Initialize session and S3 client
sess = sagemaker.Session()
bucket = sess.default_bucket()  # Get default S3 bucket for the session
role = sagemaker.get_execution_role()
region = boto3.Session().region_name
account_id = boto3.client("sts").get_caller_identity().get("Account")

s3 = boto3.client('s3')

# Define the S3 path to the file you want to download
s3_private_path_csv = "s3://{}/root/AAI-540_Predictive-Maintenance-for-Pharmaceutical-Manufacturing-Equipment/resampled_predictive_maintenance_dataset.csv".format(bucket)
bucket_name = s3_private_path_csv.split('/')[2]
key = '/'.join(s3_private_path_csv.split('/')[3:])

try:
    # Download the dataset from S3
    response = s3.get_object(Bucket=bucket_name, Key=key)
    df = pd.read_csv(io.StringIO(response['Body'].read().decode('utf-8')))

    # Print the head of the dataset to confirm
    print(df.head())

    # Save the updated DataFrame to a CSV file (if any changes are made)
    df.to_csv('resampled_predictive_maintenance_dataset.csv', index=False)

    # Define the S3 path to the updated dataset you want to upload
    s3_private_path_tsv = "s3://{}/root/aai-540-homework/homework-2-1/data/dataset.csv".format(bucket)

    # Upload the updated dataset to S3
    s3.upload_file('resampled_predictive_maintenance_dataset.csv', bucket_name, 'root/aai-540-homework/homework-2-1/data/dataset.csv')
    print(f"Dataset successfully uploaded to {s3_private_path_csv}")

except Exception as e:
    print(f"[ERROR] {e}")
